Tidy up the zh-CHS translation for "irank" - the translators apparently looked at the comment in   <data name="TransitionTreeNode_GetLabel_irank__0__" xml:space="preserve">     <value>irank {0}</value>     <comment>Isotope rank for transition node</comment>   </data>
and came up with
    <value>针对离子对节点的同位素排名 {0}</value>

which is literally "Isotope rank for transition node {0}" where what we really want is
    <value>i排名 {0}</value>

which is literally "irank {0}" (note the Roman "i" in "i排名")

One question in my mind - in the Japanese version of this, there is no blank between "irank" and the value. Is that also appropriate in Chinese?