Fixes a medium size annoyance for developers using parallel test in Skyline Tester
I (bspratt) often realize that I've kicked of the wrong selection of tests. So, Stop the tests, make the selection, then either wait for (or manually kill) the workers so a new set can be spun up. Or forget to wait and get a bunch of errors.

With this change, each round of parallel tests includes a timestamp in worker names so this conflict never happens.