Improved support for working with paths that contain spaces, or legal but unusual characters like caret and ampersand, and for the issue of stray files being left behind by various tests and 3rd party tools
Made changes to the Skyline automated testing scheme to improve our support for users with windows usernames containing spaces an other legal but unusual characters. This is achieved by setting the TMP environment variable to point at a directory whose name contains such characters. Normally that value is based on the user name, so even on a system with username "ihavenospacesinmyname" we can test running with a TMP dir with a name that contains space, caret, and ampersand.

We now properly escape quotes when passing arguments to msconvert's diaUmpire filter, as it requires nested quotes.

Additional cleanup for Crux pepXML output that doesn't properly escape the & in filenames

Added a custom attribute for tests that don't do well with unicode paths, usually due to mz5 or 3rd party tool usage. Tests without the NoUnicodeTesting attribute will run with TMP and results directories whose names include unicode characters.

Check for cleanup of our temp directory on test exit.

Explicitly clean up MSAmanda's tempfiles, which are currently left on purpose but which violate the general expectations of test cleanup. Added a couple of TODO comments about how we need to think more about MSAmanda leaves tempfiles around.

ProcessRunner.Run now accepts an optional argument "forceTempfilesCleanup" which, when true, runs the process with an altered environment in which TMP is a subdirectory of the original temp directory. Afterward the directory is deleted.

Added pwiz.Common.SystemUtil.TempDir class to make it easier to clean up after things like ExcelReaderFactory that leave tempfiles behind.

Custom tools that write to the immediate window weren't cleaning up their temp files.

Do a better job of cleaning up tempfiles when an MSFragger search is interrupted.