Fix a race condition that only shows up on TeamCity code coverage build
Fixed by not removing temp files associated with the Skyline Immediate window until that window is Disposed. Originally we didn't try to delete them at all, my recent commit deleted them when the tool writing to the window exited but that apparently introduced a race condition. So now they will accumulate during a Skyline run but should not remain afterward, which is still an improvement.