Improved handling of small molecule transition lists where document settings can help with precursor isotopes
Background: For small molecules, Skyline formerly constructed a document with precisely the contents of the transition list. But the new transitions were created with Auto-select enabled, so it was possible for a document settings change to unexpectedly alter the document contents. The most prominent effect was the creation of isotope precursor transitions for molecules with known chemical formulas - even if the triggering settings change had nothing to do with isotope peak counts.  Often this was precisely what was wanted, but it was an awkward and unintuitive extra step to tweak a random setting in order to create the isotope transitions.

Now when importing a small molecule transition list, Skyline will check to see applying if the current document settings would result in a set of transitions different from those explicitly given in the lest. If so, Skyline offers to enable Auto-select. If not, the new transitions are quietly created with Auto-select disabled.

Note that small molecule fragments are largely unaffected by Auto-select since, unlike peptides, Skyline does not have fragment prediction for generalized molecules. Filters such as instrument m/z range may have an effect, though.