TestPrositOutputToEncyclopediaLibraries was failing on parallel clients due to very small numerical differences in resulting TSV files… (#2637)
TestPrositOutputToEncyclopediaLibraries was failing on parallel clients due to very small numerical differences in the resulting TSV files. Added a match tolerance of 0.0000000001, and made the tolerance check able to handle doubles with "." or "," decimal separators. It's not clear why this happens on parallel clients, and it seems it only happens on my machine, but the differences really just come down to tiny variations in formatting doubles with decimal places well beyond any real world measurements.

Also added some quotes around path arguments in TestRunner's Docker invocation.