As part of the assay library work for small libraries I became increasingly aware of a major behavior difference in handling small molecule vs peptide transition lists.
For small molecules, if Skyline thinks it knows all the headers it just takes in the list with no confirmation step (i.e. no column picker dialog). Nice when it works, but if the user actually wanted to, say, ignore a column or deal with a misidentified column, there was no easy way to do that. Peptides, on the other hand, always present the column picker.

Now we always present the column picker dialog regardless of input type.