change build-apps.bat to default to optimization=speed, since "space" seems to cause trouble with Agilent DLLs
Without this change, building with build_apps.bat on a current compiler we get a crash loading pwiz_data_cli.dll and a message about "attempting managed execution inside a DLLMain or image initialization function".

This only started happening when I updated Visual Studio to v17.7.1, and still happens with 17.7.3. Version 17.4.4 doesn't seem to have this problem.

The issue seems to center on the use of the Agilent DLLs, for reasons unknown.  TestRunner built with optimization=space goes down hard when AgilentFileTypeTest is run. And it's apparent that the issue has been lurking for some time since there's a longstanding bit of jamfile code to carefully avoid building an Agilent related test "SpectrumList_3D_Test" when any kind of optimization was enabled. Now we only skip that if optimization=space.