Fix Audit Log error when copy/pasting protein groups (#2758)
Fixed unhandled error copying and pasting protein groups (reported by Philip)

Fix "LogException" when copying and pasting protein groups.
The audit log code was choking on the fact that ProteinGroupMetadata overrides the "Name" property which is marked with the "[Track]" attribute. The audit log code was calling using the accessor from the subclass instead of the accessor from the base class.