Use MassLynxSDK to identify lockmass function
Code assumed that lockmass function in SONAR/HDMSe data would not be multidimensional which is not correct.
This lead to it trying to read the lockmass function of the SONAR test data as if it was not multidimensional resulting in a crash.
This didn't break before because IsLockMassFunction(...) was incorrectly returning False for the lockmass function