Make it easier for developers to enable Prosit on local builds
As before, when the bjam commandline includes "--official" the Skyline build will fail if "Model/Prosit/Config/PrositConfig_production.xml" is not present for copying to PrositConfig.xml for embedding in the exe.

In this new behavior, if a file "Model/Prosit/Config/PrositConfig_production.xml" is present, the Skyline Jamfile will always copy it to PrositConfig.xml regardless of the "--official" flag.  If PrositConfig_production.xml is not present, the file "Model/Prosit/Config/PrositConfig_development.xml" is copied and embedded - it has no certificate and Prosit (and related tests) is disabled.

Also, remove a stale reference to Model\Prosit\Config\UpdatePrositConfiguration.bat in the .csproj file.

The private developer wiki has been updated to reflect this change.