Library Explorer improvements: * Peptides and/or molecules are shown in natural sort order (i.e. aware of numerical values embedded in strings) * We now show the number of things that match the filter rather than always just giving the total count * We remember the user's search string for each filter type as they choose and return to different ones * Fixed a bottleneck that was preventing display of ion mobility values
This required a new natural sort class to complement the existing one, which turns out to be filename specific. It doesn't understand that "a123.456" comes before "a123.4561" because it doesn't see "123.456" as a decimal value - it sees it "123" and "456" as two different numbers. No real surprise since it's just a wrapper for a windows call for sorting filenames.

This was all done in the service of the Feature Finding work but stands alone nicely so I've given it a separate PR.