Library Explorer improvements inspired by larger small molecules libraries: (#2885)
* Peptides and/or molecules are shown in natural sort order (i.e. aware of numerical values embedded in strings)
* We now show the number of things that match the filter rather than always just giving the total count
* We remember the user's search string for each filter type as they choose and return to different ones
* Fixed a bottleneck that was preventing display of ion mobility values

This required a new natural sort class to complement the existing one, which turns out to be only appropriate for filenames. (The existing doesn't understand that "a123.456" comes before "a123.4561" because it doesn't see "123.456" as a decimal value - it sees "123" and "456" as two different numbers.)

Not reported by any user - this was all done in the service of the Feature Finding work, which results in molecules with names like "mass123.456_RT12.3" where a natural sort is needed.  It stands alone nicely so I've given it a separate PR.