Perform fewer pairwise retention time alignments (#2848)
Sped up "Aligning retention times" by only performing alignments to first replicate in document (requested by Mike).

This change makes it so that Skyline no longer aligns all replicates against all of the things with retention times (i.e. runs in the libraries).
Instead, everything is aligned against the first replicate in the document (actually, the first internal standard in the document, or, if there are no internal standards then the first replicate).
Also, if "Batch Name" has been specified for replicates, then the replicates are also aligned against the first replicate within each batch (not sure if that makes a difference).