Fix some small molecule tests which have been failing when run individually from TestRunner.
TestRunner sets "Program.FunctionalTest" to "true" at the beginning of the test run, but that typically gets set back to false by "MyTestCleanup".
When Program.FunctionalTest is true, audit logging was getting turned on when saving a document as part of "convert to small molecules" which then, because of PR #2919, the document gets assigned a new GUID which causes an Assert.AreEqual to fail comparing the settings of a round-trip document.