Feature detection via Hardklör integration with Skyline (#2495)
Added a new menu item "Files > Import... >Feature Detection..." which invokes a workflow employing a lightly modified version of Hardklor/Bullseye to detect peptide-like features in mass spec data.

The workflow runs these steps on selected replicates:
  Run Hardklor and Bullseye on each replicate
  Perform postprocessing on those results to unify features appearing in multiple replicates
  Generate a BiblioSpec library with the unified features
  Add the contents of the library to the Skyline document as new targets
  Perform chromatogram extraction on the replicates with the new feature targets.

Most of the modifications to Hardklor have to do exposing the Averagine chemical formula and mass shift that Hardklor uses to generate the isotope envelope for each reported detection. Normally Hardklor discards this information and outputs just the mass and charge of a feature, but for Skyline's purposes we want enough information to construct the  the entire isotope envelope. The other significant change to Skyline's copy of Hardklor is that it won't emit matches on fewer than three isotopic peaks - it's just too noisy to be useful, otherwise.

Skyline/BiblioSpec describes the new features as small molecules with the Averagine chemical formula and mass offset produced by Hardklor, but it is important to understand that this  is not an actual identification - it's just something that produces a theoretical isotope envelope that is sufficiently similar to what's found in the data to suggest a peptide-like feature. 

Detected features are named in the library as their mass and the retention time in the replicate that yielded the best scoring peaks for the feature.

This workflow introduces a new BiblioSpec score type, "HARDKLOR IDOTP". BlibBuild is extended to read the customized Hardklor/Bullseye format, producing .blib files that contain precursor information only (this is an unusual use of the .blib format, which generally is about fragments along with precursors).

A note on "idotp" as used in Skyline vs "correlation score" as used in Hardklor:  Skyline deals with isotope envelope shape correlation as "idotp" which is a Normalized Contrast Angle (NCA). Hardklor uses Cosine Angle Correlation scoring, this is related to Normalized Contrast Angle as NCA=1.0-(acos(CAC)*2/PI). In the Skyline Feature Detection user interface, and in BiblioSpec, we use idotp (NCA), and translate back and forth as needed for Hardklor input and output.

To retain Hardklor's sense chemical formula and mass shift, we use Skyline's extensions to chemical formula descriptions as used in crosslinker descriptions - they may contain information about unexplained mass shifts, e.g. "C12H5O3[+1.2]" (unexplained additional mass of 1.2 AMU), "C12H5O3[-1.3]" (unexplained reduced mass of 1.3 AMU), "C12H5O3[+1.2/1.21]" (unexplained additional mass of 1.2 mono / 1.21 average) etc. The + or - sign is required, the value is always Invariant formatted ("." decimal separated) and all must be enclosed in square brackets..

---------

Co-authored-by: Brendan MacLean <brendanx67@users.noreply.github.com>
Co-authored-by: nickshulman <nicksh@uw.edu>