Fix incorrect peak areas with very fast cycle time Fixed peak area sometimes incorrect when cycle time approaches 1000 per minute (reported by Joshua)
"EvenlySpaceTimes" does not bother to interpolate if the variance between cycle times is less than 1000/min.
This means that times might not be evenly spaced, but ChromPeak constructor was assuming that the points were evenly spaced, and so multiplied by the difference between the first and second time points to convert from Crawdad sum of peak heights to area.
This change makes it so that it multiplies by total peak width and divides by number of points.