EditRemoteAccountDlg: Set "Test" button text to "Connect" when applicable.
Skyline/Properties/Resources.resx has been updated with the "Test" and "Connect" strings to support Internationalization.

The default label for the button in the Designer is "Test/Cnnct" so that users of the Designer will know it supports both values.
This may need to be changed since it adds a string to be internationalized that does not get shown to the user.