AppCodeCheck:Handle Status 404 when AppCode Not Found in Ardia
When the ApplicationCode in Skyline is not in Ardia, besides the response body the HTTP status code is 404.
So when get HTTP status code 404 need to process the response body to see if it is the standard response JSON and if it is then the response body values are evaluated, otherwise it is reported to the user that the URL was not found.