added static code inspection for values that don't agree between Settings.settings and app.config
a little bit of voodoo around settings that are collections of strings - Settings Designer may not create an app.config entry if the collection is empty. On the other hand, if the collection is non-empty then edited to be empty, the app.config entry will exist, but with an empty string as its value.