Improved support for metal adducts (for PFAS research) and more complex adducts seen in popular CCS libraries (#3201)
At the Asilomar 2024 ion mobility conference it was clear that there's a lot of interest in PFAS research and CCS libraries in general. This commit improves Skyline support for metal adducts which are common in PFAS, and other novel-to-us adducts seen in popular libraries like Maclean lab's CCS Compendium and Erin Baker lab's Lipid and PFAS databases.