Fix duplicate entries on Calibration Curve legend Fixed duplicate entries on Calibration Curve legend when first shown (reported by Brendan)
The line "this.calibrationGraphControl1.Options" was not supposed to be in CalibrationForm.Designer.cs, and was causing "Settings_OnPropertyChanged" to get called in the middle of "Update" which resulted in two sets of curves being added.
Also, added "[Browsable(false)]" to a few properties that were not supposed to be seen in the property sheet.
And changed a few "[EditorBrowsable(EditorBrowsableState.Never)]" to "[Browsable(false)]" because "EditorBrowsable" was the wrong attribute to be using to hide something from the property sheet.
"EditorBrowsable" is use for IntelliSense.