Tweaks to deal with new warnings from latest ReSharper.
This one is very popular:
https://www.jetbrains.com/help/resharper/UsageOfDefaultStructEquality.html
It suggests that default runtime comparisons are probably slow, and we should use our own code for equality checks. I've done so, butI worry a bit that we now have more code to maintain in the event that a member is added to a struct. (Interestingly in the act of doing this, I noticed that Model.TransitionGroup.LossId already has its own equality comparer, but doesn't include all its members in that - I've left a CONSIDER comment there)