Enable running perftests via Test Explorer in Visual Studio
This change makes it so the default value of "RunPerfTests" is true.
This means that if you run the test in Visual Studio using either Test Explorer or ReSharper's Unit Test Explorer, the test will actually get run.
This change does not affect TestRunner (or SkylineTester) because TestRunner always explicitly sets RunPerfTests to false if it has not been specified on the command line.

Also, this change makes the default for "RunSmallMoleculeTestVersions" true as well.