Simplified handling of bogus lockmass correction requests
While chasing down memory issues, I noticed that there seemed to be some proximity to tests that involved asking for lockmass correction on data that isn't actual Waters raw. I'm not 100%% convinced that this is the issue, but it does seem just as well to simplify this code so that the data processing chain is left completely untouched when spurious requests are made. If nothing else it makes for a more accurate set of CVParams.

not reported by any user