Bjam build targets for "pwiz_tools\Skyline\DevTools\ResourcesOrganizer" (#3347)
quickbuild.bat pwiz_tools\Skyline\Executables\DevTools\ResourcesOrganizer//IncrementalUpdateResxFiles
This command should be used periodically during the development cycle when the positions of controls on forms have changed, but the Japanese and Chinese .resx files are out of date so the localized controls are in the wrong places.
It will revert to the English all strings that have changed since last version.
Last version's strings are in the file "pwiz_tools/Skyline/Translation/LastReleaseResources.db"

quickbuild.bat pwiz_tools\Skyline\Executables\DevTools\ResourcesOrganizer//FinalizeResxFiles
This command should be used after visual freeze. It will update the .resx files and add <comment> tags for everything that will be sent to localizers.