Reduce memory usage of "ExplicitPeakBounds" and "SpectrumMetadata" (#3374)
Add some subclasses of "ImmutableList" which can store some types of data more efficiently:
ConstantList: for when all of the elements in the list are identical
IntegerList: for when the elements are integers but their values are in a range that can be represented by less than 4 bytes.
Factor: for when the list is long but the number of unique values is small so they can be represented as integer indexes into the list of unique values
NullableList: stored as a list of struct's alongside a list of bits.