Skyline: Raised the peak truncation threshold (#3361)
The TransitionResults.Truncated and PrecursorResults.CountTruncated values have been reported as too sensitive for a while, especially for the Stellar which is often tuned to use very narrow retention time windows for scheduled data acquisition.
These values are based on a height proportion threshold currently set to 0.01 or 1%.
In the initial commit the value is set to 0.25 of 25%, i.e. quarter-max where 50% is half-max.

Added "Proportion Truncated" report field