Fix exception in AxisLabelScaler (#3435) (#3439)
Fix exception in AxisLabelScaler (#3435)

Fixed error sometimes trying to draw text on graphs.

There was a bug in the implementation of the setter for FontSpec.Size where "_scaledSize" would be a nonsensical number which grew larger every time the size was changed and would eventually overflow.

Fix incorrect using from previous commit to release branch