Fix bug with modified sequence containing "F[+10]"
Fixed bug where "+10" was incorrectly interpreted as "Label:13C(6)15N(4) (R)" regardless of which amino acid it was applied to (reported by Philip)

Changed "SimplifyUnimodSequence" to ignore mismatches between unimod modification symbols and the amino acid they could be applied to.
Changed to catch exceptions in MassListImporter.DoImport and display an error with line number.