Re-order the actions taken when saving a Skyline document.
This is a rudimentary change and not ship-ready. Rather, it is option #1 of two ways to get FilesTree the information it needs to respond to Save As actions. It will be replaced shortly with option #2 (firing a new DocumentSaved event) for a head-to-head comparison of these options.

FilesTree needs the new path *and* successful writes of all files (.sky, .view, .skyl, etc) to the new location to correctly respond to Save As actions. Without this informtion, FilesTree shows "missing" icons for files not yet written to disk, puts "file missing" messages in tooltips, etc and remains in this state until the next DocumentChange event.

What changed?

The SetDocument(new, orig) call becomes the last step in the sequence rather than the first.

This causes the DocumentChange event raised when saving the doc to include the new path and fire after files are successfully written to disk.

Despite being a low-level change, only 1 test broke - TestChangeDocumentGuid on line 76 because the newly saved doc's GUID isn't updated as expected.