Add a DocumentSaved event to SkylineWindow and use it to update FilesTree in response to a Save As action.
This event fires for all Save actions and includes a bool used to distinguish Save As actions that write to a new path.

FilesTree subscribes to this event and updates itself when IsSaveAs = true giving the tree access to the new path and all files successfully written to disk.

As mentioned in the previous commit (c39c101), this is option #2 for updating FilesTree in response to Save As actions.

This option is preferred over Option #1 because:
* reduces the async work used during FileTree event handling
* reuses logic in MergeNodes for initializing a tree node's FileState - 1 code path > 2 code paths.
* does not require outsatnding, complex re-work of SkylineFiles.SaveDocument(string, bool, bool) needed to make Option #1 production quality

This change also reverts the previous commit implementing Option #1.