Remove unneeded progressStatus.NextSegment. Fix indentation. Change AlphaPeptDeepLibraryBuilder to use CPU when running in simulation mode NONVIDIAHARD.
When running AlphapeptdeepBuildLibraryTest, run in simulation mode NONVIDIAHARD to prevent the test node from trying to install and use NVIDIA gpu (which requires admin elevation). Attempting this to fix Amazon test node failure, when NVIDIA card is present but we don't want to install any NVIDIA software.

Allow test to deletePython or keep it for faster testing.