Pass correct value for "IgnoreZeroIntensityPoints" when opening MSDataFile
Fixed "error reading spectrum" extracting chromatograms from certain data files (reported by Nikki)
Skyline used to pass "true" for "IgnoreZeroIntensityPoints", and that got changed to "false" which exposes a "Index was outside the bounds of the array" error in "Agilent::FrameImpl::getCombinedSpectrumData".
There might be a different bug in "getCombinedSpectrumData" which we should also fix.