Add a new files section in a Skyline Document that will contain references to user-defined files. For example: standard operating procedures, spreadsheets, abtitrary PDFs, etc.
This checkpoints early work and moves it out of the FilesView branch into what will be a separate PR so the main PR can be merged and available in Skyline dailies soon.

Key to this change are new Immutables for files / file with XML serialization, find-by-ID, and basic unit tests.

This work is far complete - all names and structure are subject to change.

Tests: FilesTreeFormTest passes, including tests for the new section in SrmSettings. Other tests may fail.