Ardia File Upload (#3463)
Add support for uploading the current Skyline document to the Ardia Platform via a new menu item under File => Upload to Ardia that appears when an Ardia account is registered with Skyline. 

Includes:
1) New dialog for choosing the destination folder for the .sky.zip file on the Ardia server (PublishDocumentDlgArdia). UI looks similar to the dialog for uploading to Panorama.
2) Uses ShareTypeDlg to configure files to include in the .sky.zip file
3) Supports uploading single and multi-part documents to the Ardia server

Also includes new approach for calling remote APIs that isolates the network layer from the application and UI layers (ArdiaClient).

Co-authored-by: Aaron Banse <acbanse@icloud.com>