Remove "default" option from run-to-run alignment dropdown (#3608)
People were getting confused about the difference between the "default" option, and the explicit option that was the same as the default.
The explicit option that was the same as the default has been removed.
Added a tooltip explaining what all the options do and which regression method they use.
Also, show replicate name on Y axis in Score to Run regression graph when single replicate is chosen.