More graceful handling of libraries containing nonsensical entries (e.g. water loss on molecule containing no oxygen atoms)
Formerly the Library Explorer would throw an exception when presented with, for example, a water loss adduct on a molecule with no oxygen in it. Now informative error messages are shown in the Immediate Window for such entries when the library is first loaded, and in the Explorer window for individual entries as they are being inspected. (Ideally such things wouldn't be in the library in the first place, but we don't always have control over how libraries are created)

Note that molecules described by mass only are permitted to have losses so long as the resulting mass is greater than zero, but for molecules supplied with formulas we need a sensible resulting formula to create isotope envelopes etc.

Reported anonymously via ExceptionWeb