Fix intermittent failure in TestMultiInjectRescore (#3660)
The test called "Close" on the AllChromatogramsGraph
The AllChromatogramsGraph does not have a close button, and if it is closed in this way SkylineWindow will not know that it has been closed and will throw exceptions if it tries to send it update that window.
Changed the test to not close that window.