Improved handling of Unicode in file paths
Many external tools, and some vendor reader APIs, can't deal with Unicode characters in file paths. This is often troublesome for users on non-English-language systems. This PR solves that problem by selective use of Windows 8.3 short filename conversion when it is available.

e.g. for the mz5 reader, which can't deal with unicode in paths, we try to use the Windows API to change
   "C:\Program Files\Common Files\my files with ünicode\foo.mz5"
(note the umlaut U) to
   "C:\Program Files\Common Files\MYFILE~1\foo.mz5"
before passing the path into the reader API.

This only works on NTFS volumes, and only those with 8.3 support enabled. So, not on Docker instances, or under Wine, or on NTFS volumes with 8.3 short filename support disabled.

Conversions are only attempted where required - any external tool or vendor reader that is known to handle unicode paths won't get this treatment applied to any file paths we hand to it. That would be unnecessarily confusing.

We already have code for testing Skyline with Unicode paths, and a test attribute "NoUnicodeTesting" to skip that for tests of Skyline features known to have trouble with Unicode - usually because they spin up processes for tools that can't deal with Unicode paths. With this PR that test attribute is not currently in use - there are no tests that fail when Unicode paths are involved. The Skyline code now converts file paths to 8.3 as required by each external tool before launch. The pwiz wrappers for vendor readers do 8.3 conversion where needed, too.

Obviously we don't test Unicode path handling in environments where we know 8.3 conversion isn't available.

And obviously a user running Skyline on a system without 8.3 short filename support is still going to experience difficulty with Unicode paths as before. But this should help in the majority of cases.