Improved handling of Unicode in file paths
Many external tools called by Skyline, and some mass spec data readers used in Skyline (and other ProteoWizard tools like msconvert), can't deal with Unicode characters in file paths. This is often troublesome for international users. We now solve that problem by selective use of Windows 8.3 short filename conversion when it is available.

For example the mz5 reader library can't deal with Unicode in paths, so we ask Windows for the short name of the file path (note the umlaut U) 
   "C:\Users\Günter\data\run1.mz5"
which is something like
   "C:\Users\GUNTER~1\data\run1.mz5"
and we pass that to the mz5 reader, which can then read the file with no problems.

This works equally well for non-European languages, but it only works on NTFS volumes, and only those with 8.3 short filename support enabled. So, not on Docker instances, or under Wine, or on NTFS volumes with 8.3 short filename support disabled.

Conversions are only attempted where required - any external tool or file reader that is known to handle Unicode paths won't get this treatment. That would be unnecessarily confusing.

Not requested by any user, but the need was observed by BrianP at a Skyline training in Germany.


Developer details:

We already have code for testing Skyline with Unicode paths, and a test attribute "NoUnicodeTesting" to skip that for tests of Skyline features known to have trouble with Unicode - usually because they spin up processes for tools that can't deal with Unicode paths. With this PR that test attribute is not currently in use - there are no tests that fail when Unicode paths are involved. The Skyline code now converts file paths to 8.3 as required by each external tool before launch. The pwiz wrappers for vendor readers do 8.3 conversion where needed, too.

Obviously we don't test Unicode path handling in environments where we know 8.3 conversion isn't available.

And obviously a user running Skyline on a system without 8.3 short filename support is still going to experience difficulty with Unicode paths as before. But this should help in the majority of cases.