Avoid potential test issues by not calling Invoke() in SkylineInvoke() if already on the UI thread.
When I attached a debugger to a hung test it was found to be in a nested Invoke() call, which seemed suspicious to me. Many of our less stable tests were subsequently shown to have nested Invoke() calls due to RunUI() blocks calling RunUI() further down the stack.

Calls to Invoke() on the UI thread are not technically wrong, but according to ChatGPT they "can easily lead to deadlocks, re-entrancy bugs, and stack blowups, depending on the framework and context". A simple check for InvokeRequired avoids potential issues.