Fix TestPerfTicArea
My change in PR #3645 to SpectraChromDataProvider.GetChromatogram mistakenly removed the code that calculates the Total Ion Current area in the case that the area had to be calculated from all of the spectra. The TIC area was only being calculated in the case that the TIC chromatogram could be read from the raw file and had enough information to filter out non-MS1 spectra.